+++ test/migrations/change_column.rb
@@ -437,12 +437,8 @@ class ChangeColumnAlgorithmInplace < TestMigration
 end
 
 class ChangeColumnLockShared < TestMigration
-  def up
+  def change
     change_column :users, :country, :string, limit: 21, lock: :shared
-  end
-
-  def down
-    change_column :users, :country, :string, limit: 20
   end
 end